fix: 合并智谱渠道 (*获取模型报错，但API实测可调用)
