+++ relay/channel/zhipu/constants.go
@@ -17,7 +17,7 @@
 package zhipu
 
 var ModelList = []string{
-	"chatglm_turbo", "chatglm_pro", "chatglm_std", "chatglm_lite",
+	"glm-4", "glm-4v", "glm-4-plus", "glm-4-air", "glm-4-airx", "glm-4-flash", "glm-4v-plus", "glm-z1-flash", "glm-4.5", "glm-4.5-flash", "glm-4.5-X",
 }
 
 var ChannelName = "zhipu"
