+++ relay/channel/zhipu/dto.go
@@ -21,42 +21,55 @@ import (
 	"veloera/dto"
 )
 
-type ZhipuMessage struct {
-	Role    string `json:"role"`
-	Content string `json:"content"`
-}
-
-type ZhipuRequest struct {
-	Prompt      []ZhipuMessage `json:"prompt"`
-	Temperature *float64       `json:"temperature,omitempty"`
-	TopP        float64        `json:"top_p,omitempty"`
-	RequestId   string         `json:"request_id,omitempty"`
-	Incremental bool           `json:"incremental,omitempty"`
-}
-
-type ZhipuResponseData struct {
-	TaskId     string         `json:"task_id"`
-	RequestId  string         `json:"request_id"`
-	TaskStatus string         `json:"task_status"`
-	Choices    []ZhipuMessage `json:"choices"`
-	dto.Usage  `json:"usage"`
-}
-
+//	type ZhipuMessage struct {
+//		Role       string `json:"role,omitempty"`
+//		Content    string `json:"content,omitempty"`
+//		ToolCalls  any    `json:"tool_calls,omitempty"`
+//		ToolCallId any    `json:"tool_call_id,omitempty"`
+//	}
+//
+//	type ZhipuRequest struct {
+//		Model       string         `json:"model"`
+//		Stream      bool           `json:"stream,omitempty"`
+//		Messages    []ZhipuMessage `json:"messages"`
+//		Temperature float64        `json:"temperature,omitempty"`
+//		TopP        float64        `json:"top_p,omitempty"`
+//		MaxTokens   int            `json:"max_tokens,omitempty"`
+//		Stop        []string       `json:"stop,omitempty"`
+//		RequestId   string         `json:"request_id,omitempty"`
+//		Tools       any            `json:"tools,omitempty"`
+//		ToolChoice  any            `json:"tool_choice,omitempty"`
+//	}
+//
+//	type ZhipuTextResponseChoice struct {
+//		Index        int `json:"index"`
+//		ZhipuMessage `json:"message"`
+//		FinishReason string `json:"finish_reason"`
+//	}
 type ZhipuResponse struct {
-	Code    int               `json:"code"`
-	Msg     string            `json:"msg"`
-	Success bool              `json:"success"`
-	Data    ZhipuResponseData `json:"data"`
+	Id                  string                         `json:"id"`
+	Created             int64                          `json:"created"`
+	Model               string                         `json:"model"`
+	TextResponseChoices []dto.OpenAITextResponseChoice `json:"choices"`
+	Usage               dto.Usage                      `json:"usage"`
+	Error               dto.OpenAIError                `json:"error"`
 }
 
-type ZhipuStreamMetaResponse struct {
-	RequestId  string `json:"request_id"`
-	TaskId     string `json:"task_id"`
-	TaskStatus string `json:"task_status"`
-	dto.Usage  `json:"usage"`
+//
+//type ZhipuStreamResponseChoice struct {
+//	Index        int          `json:"index,omitempty"`
+//	Delta        ZhipuMessage `json:"delta"`
+//	FinishReason *string      `json:"finish_reason,omitempty"`
+//}
+
+type ZhipuStreamResponse struct {
+	Id      string                                    `json:"id"`
+	Created int64                                     `json:"created"`
+	Choices []dto.ChatCompletionsStreamResponseChoice `json:"choices"`
+	Usage   dto.Usage                                 `json:"usage"`
 }
 
-type zhipuTokenData struct {
+type tokenData struct {
 	Token      string
 	ExpiryTime time.Time
 }
